dumpstate_test: Fix failing tests
With Ibcef6b107de4327f769242c0c43d7f5317eee7f6, we introduced
logic that would avoid outputting the duration if it was shorter
than 500ms.  As a result, some of these tests fail when they
take less than 500ms, because they were expecting the duration
to be output regardless.

We change the tests to no longer look for the duration in the
output.

Test: dumpstate_test run now has every test pass
Bug: 133225555
Change-Id: I6d332e61bf6ac6832459e896b4a30436119930e9