Update patch set 3
Patch Set 3:

Answering my own question, this is working because the version of external/mesa32 on pie-gsi doesn't have a build target for libGLES_mesa, so nothing gets installed on the image. As soon as we upgrade the library we'll have multiple GPU implementations, and will run into the bug that this change corrects.

Patch-set: 3
Reviewer: Gerrit User 1062926 <1062926@85c56323-6fa9-3386-8a01-6480fb634889>