Binder: support storing interface stability
An internal CL adds stability annotations to interfaces. In order to
make it easier to make changes against this on AOSP, the API only is
merged here. This means, until Android Q is released, AOSP will have
stability API annotations but not enforcement. Enforcement will only
happen on internal branches. This is being done because of a mismatch
between vndservicemanager (in the vendor.img blobs) and libbinder in the
VNDK.

Bug: 136027762
Test: TH

Merged-In: Ia637ee3652d55550e7fce78876458f391b1dd928
Change-Id: Ia637ee3652d55550e7fce78876458f391b1dd928