Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (1 comment)

I agree with the flow you suggested. I happened to assume that we are exiting the process when we log an error(MYLOGE). This was my mistake. So from what I assumed this CL will exit the process
1. If acquire wake lock fails
2. If atexit fails
3. If release wake lock fails 
Since we don't exit process for any of these, I have the same question as you, should we continue BR in all of these three cases?

Patch-set: 3
Reviewer: Gerrit User 1530043 <1530043@85c56323-6fa9-3386-8a01-6480fb634889>