Update patch set 4
Patch Set 4:

Read the issuetracker. It seems to be a driver/HAL implementation issue. Regardless of number of hardware device that backs wakeup and non-wakeup version sensor in HAL, the HAL implementation should do as much as it can to guarantee independent operation of two sensors.

A fix in driver will be preferred.

Patch-set: 4
Reviewer: Gerrit User 1074523 <1074523@85c56323-6fa9-3386-8a01-6480fb634889>