Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> any opinion on if this is a good idea? seems it was an interesting approach but maybe not the best

> Patch Set 2:
> 
> any opinion on if this is a good idea? seems it was an interesting approach but maybe not the best

I like this because this is more correct. But I also agree with you that the path logic might be complex and it can impact performance.

Can we use both of the two approaches? For example, we could do the dynamic setting only when __ANDROID_VNDK__.

Patch-set: 2
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>