surfaceflinger: fix default frame/viewport assignment
At bootup, when the DisplayManager isn't started, surfaceflinger sets
the default frame and viewport to the display size.

However, when a default orientation is set to 90 or 270, the axis need
to be swapped, otherwise the frame/viewport is in the wrong direction,
ending up cropping the bootanimation to the display physical height.

Change-Id: Ib1d6dc85e71dc695562755348027363a5813620f
Signed-off-by: Gary Bisson <gbisson@essential.com>