Revert the change which switch the handle to holding a strong pointer
Layer will not be removed if the Handler's BpBinder has the reference
to the  BBinder, since it is a strong reference to the Layer,
even though we have called the SurfaceControl's remove on the Client side,
the Layer will not be recycled. So lots of Layers will take up a lot of memory,
and the deivce runs slowly and hang.

Bug: 140880272
Test: multi-click split screen control button(More and more clicks)

Change-Id: I2822edae945f1274d64ec198720ee810f9f5f3be
Signed-off-by: MiaoJian <miaojian1993@gmail.com>