Update patch set 2
Patch Set 2:

Hi Lukasz,

> Patch Set 2:
> 
> Hi Gary, do you happen to have any other issues related to 270 degree rotation? I'm working on such device myself and I'm experiencing a lot of other issues. For example it looks like only some layers are being rotated (statusbar seems fine but everything else is wrong), perhaps only client or hwc composition is being rotated properly.

Yes I noticed other issues, especially towards animation (rotation anim is the worst example).

But note that this patch is only about bootanimation, when surfaceflinger sets the default frame/viewport on its own.

> I kinda hacked around it with your patch + https://gist.github.com/luk1337/234dbb7fd951650a3f941dbc5dcba103 but now it looks like client composition is broken cause all animations are corrupted...
> 
> Is there some patch that I may have not noticed on Gerrit that'd magically fix all issues that I'm having?

On P we had everything working properly with this patch for bootanimation and this one for all animation problems:
https://android-review.googlesource.com/c/platform/frameworks/native/+/766681

Unfortunately this patch doesn't apply on Q and isn't an ideal solution as it relies solely on the CaptureFill differentiation.

I'm trying to get some traction on that bug to have someone acknowledge the issues when using 90/270 install orientation.

Patch-set: 2
Reviewer: Gerrit User 1406250 <1406250@85c56323-6fa9-3386-8a01-6480fb634889>