Update patch set 1
Patch Set 1:

For context: b/139096173 describes the issue. The Java "createFromParcel()" method uses the caller-provided size of a Parcelable to advance the data position, which allows a caller to move the data position to an unaligned address, which can result in a crash on some architectures.

Solutions I thought of were:
1) Check the passed in parcelable size is a multiple of 4 in the auto-generated code
2) Modify the auto-generated code to calculate the correct offset itself, based on the types it reads (eg, reading an int adds 4 bytes, etc.), and either use that size and reject it if it's different from the user passed-in size.

I rejected 1) because it would result in a code increase in all our Parcelables, which is not desirable, and 2) because it's a pretty large change to the AIDL generator (if feasible at all).

This change instead is pretty local. We should probably print a warning when setDataPosition() is called with a number that isn't a multiple of 4. It doesn't happen on my device.

Let me know if you have other ideas!

Patch-set: 1