[DO NOT MERGE] Sync visible regions on  winscope traces for SF
Currently the Winscope traces for SF don't have the correct visible
regions, making them hard to debug. This happens because the drawing
state is updated on the message "Invalidate" and the visitble regions
are calculated during the message "Refresh".

We now use a recursive mutex to block the tracing while those messages
are processed. Moreover, we start tracing as soon as the visible regions
are calculated, instead of waiting the frame to be posted, allowing the
trace to run in parallel to the frame composition

Test: Enable surfaceflinger tracing. Record a systrace. Check for
entries of type: traceLayersLocked

Change-Id: Idc6d9f464e21a98d169c622326884c8eb20f958e