[DO NOT MERGE] Start Winscope tracing after visible regions are set
Currently winscope tracing happens after the composition is finished.
Depending on the device load this means that winscope tracing may be
running in parallel to the next composition frame and may record invalid
data. With this change we start the trace as soon as the information is
available and mitigate the possibility of recording invlalid state data.

Test: Enable Winscope SF tracing. Record a systrace. Use the device.
Check that the traceLayersLocked thread is triggered after
rebuildLayerStacks, while doComposition happens

Change-Id: Id0d81883a335a7c8283d855c44ff86fc964b8315