Rename # vndk tag to # llndk
The APIs that are tagged with # vndk are actually for LLNDK libraries.
Although LLNDK is part of VNDK, calling those APIs 'vndk' has given
users a wrong perception that the APIs don't need to be kept stable
because that's the norm for most of the VNDK libraries that are not
LLNDK.

In order to eliminate the misunderstanding, rename the tag to 'llndk' so
that people introducing new such API will realize what they are signing
themselves up for.

Bug: 143765505
Test: m
Merged-In: I10d40e06b1f1d6d0e190699ab6356b18d53fd2ab
(cherry picked from commit 0a0e26b17b9fcc914e3d475d4619ef81972973e7)
Change-Id: I10d40e06b1f1d6d0e190699ab6356b18d53fd2ab