Update patch set 1
Patch Set 1:

Can you consider turning off -Weverything for these modules?  -Weverything is literally everything, adds an extra burden on the toolchain and build teams to deal with new experimental warnings in your code, and is not recommended for production code.  See for example https://quuxplusone.github.io/blog/2018/12/06/dont-use-weverything/.  -Wextra would be more appropriate.

If we can get to 0 uses of -Weverything in the tree I will forbid it in the build system.

Patch-set: 1