Update patch set 2
Patch Set 2: Code-Review+2

This should be fine. 

getTotalSize() is used by SurfaceFlinger as an int64_t in tracing (ATRACE_INT64), but if the total size is so large that that we need the highest order bit of a uint64_t, something has gone horribly wrong.

Also casting in dump from uint64_t does lose precision but it really doesn't matter in this case since the number shouldn't be large and it is only a (bad) estimate anyways.

Patch-set: 2
Reviewer: Gerrit User 1073546 <1073546@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2