dumpstate: always printf the duration of command to bugreport
Google bring change b86f33be to ignore printting the duration
of command shorter than 500ms to bugreport. However, it is not
convenient for developers to grep|sort the dumpstate sections.

I think the right thing to do is:
(1) always do the printf to bugreport.
(2) Keep the logcat logging the same:
     if (elapsed >= .5f || verbose_) {
         MYLOGD("Duration of '%s': %.2fs\n", title_.c_str(), elapsed);
      }
(3) remove logcat_only_ because I can't think of any reason where we
    want to only log to logcat.

Change-Id: I947fdd514eb31b6a6058dd841d15ba41aba063ac
Signed-off-by: chenqiwu <chenqiwu@xiaomi.com>