Update patch set 2
Patch Set 2:

> Patch Set 2: Autosubmit+1 Presubmit-Ready+1
> 
> > Patch Set 1:
> > 
> > Do you know why this is there?
> 
> Orion added these. Want him to review this as well.

There are here because the generator script splits together variables templates to make java files that have them in.

<TL;DR>

When I added them I was solving a different problem - I needed to regenerate the generated (and checked-in) source files. Unfortunately not everyone had appreciated there was a generator script so changes had been committed to the generated files. So we had to fix the generator to match the manually added contributions to the generated files.

</TL;DR>

Patch-set: 2