Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> FYI, b/146310935 appears to be why this is failing presubmit. Planning on merging this as is once that is fixed (unless you have any clean ideas for the string copy).

The only thing that comes to mind is if this is commonly a single-owner String8 then a detach() method or similar would be the easy fix. If it's not commonly a single-owner String8, though, then the copy is needed anyway and it's fine.

Patch-set: 2
Reviewer: Gerrit User 1010103 <1010103@85c56323-6fa9-3386-8a01-6480fb634889>