[Bugfix]add onBufferDetached callback to release the <sp>GraphicBuffer reference of Surface
if call acquireBuffer to acquire GraphicBuffer from BufferQueue1, and then call ImageWriter.queueInputImage to do prevOwner.detachImage, and then call attachAndQueueInputImage
to do attach and queue the GraphicBuffer to BufferQueue2,
we can only detach the BufferQueueCore's <sp>GraphicBuffer reference from BufferQueue1,
but can not release the <sp>GraphicBuffer reference that mSlots[slot].buffer in Surface from BufferQueue1 until Surface is destroyed.

In this case, the GraphicBuffer not be released causes memory redundant use.
So we add onBufferDetached callback method interface support in IProducerListener
and add releaseSlot method in Surface to release the <sp>GraphicBuffer reference.

Test: the test results:We can release 600M memory at xiaomi Camera, this issue can also test in two BufferQueues scene,as mentioned above.

Change-Id: If95c6e293dfc8cc2116f5825a118602625152c70
Signed-off-by: Wangrui <wangrui26@xiaomi.com>