[Bugfix]add onBufferDetached callback to release the <sp>GraphicBuffer reference of Surface
if call acquireBuffer to acquire GraphicBuffer from BufferQueue1, and then call ImageWriter.queueInputImage to do prevOwner.detachImage, and call attachAndQueueInputImage
to do attach and queue the GraphicBuffer to BufferQueue2,
we can only detach the BufferQueueCore's <sp>GraphicBuffer reference from BufferQueue1,
but can not release the <sp>GraphicBuffer reference that mSlots[slot].buffer in Surface from BufferQueue1 until Surface is destroyed.

And ConsumerBase has implemented the callback onBufferReleased which will be called to notify comsumer release <sp>GraphicBuffer reference in comsumer,when producer calls detachBuffer,
but Surface has no similar callback method to release the <sp>GraphicBuffer reference in Surface,
so Surface also need a similar callback method to receive the release notification when consumer do detachBuffer.

In this case, the GraphicBuffer not be released causes memory redundant use.
So we add onBufferDetached callback method interface support in IProducerListener
and add releaseSlot method in Surface to release the <sp>GraphicBuffer reference.


Test: the test results:We can release 600M memory at xiaomi Camera, this issue can also test in two BufferQueues scene,as mentioned above.

Change-Id: Ib62f87a7c82bd123fa15904b38e0cfb20b58c261
Signed-off-by: Wangrui <wangrui26@xiaomi.com>