Fix installd mount paths with FUSE
When parsing mounts from procfs in installd, we should match against
/mnt/pass_through/0/ instead of /mnt/pass_through as the former is
guaranteed to always be mounted even when multiple users are running.

Additionally we can remove the hack added in
Ic42574a46c4408ce5d43d4f412dce7f9876b8044 to work around the fact that
the pass_through mount may not be mounted in some cases and so
fallback on the lower filesystem path. This fallback caused the linked
test to fail because sdcardfs cache got out of sync when we
deleted directories from the lower fileysystem and bypassed sdcardfs.

Bug: 135341433
Test: atest android.appsecurity.cts.StorageHostTest#testCache
Change-Id: I77ee304fc2e6c3a619d94ca66cd664df7f1743d6