Remove need for libbinderthreadstate.
Instead of having this library, libbinder/libhwbinder can keep track of
stack pointers so that when they recurse, we know which one was visited
most recently.

As with the original implementation of libbinderthreadstate, this is
somewhat of a hack. An explanation of why this is and what to do instead
is addeed in CallerUtils.h.

Bug: 148692216
Test: wip (TODO: unittests as well)
Change-Id: Ief28663728fb8786b06bf9e72238052b9af81d87