Update patch set 6
Patch Set 6:

> > Kui, this code path no longer exists in our latest branch, hence we will clean this up internally and see what we can do with AOSP. Before that, we will not accept this change. We will get back to you when we have something to share.
> 
> Hi google teams, 
> 
> AOSP is open source, due to unreleased code, patch can't merge,this affects
> developer's enthusiasm for contributing code.

Hi Kui,

your concern is well received from our side. And we do appreciate your feedback and your contribution. But please expect delay on our side due to the coronavirus situation, our engineers are all overwhelmed at this point.

Regarding the merge conflict, please do kindly understand that the majority of development happens internally before we release to AOSP and hence it could have merge conflict naturally because of the development workflow.

And I also want to clarify that merge conflict is not a blocker here. When we receive a patch we then must 1) try to reproduce the issue, 2) evaluate the fix both from internal and AOSP and 3) verify it carefully before we can proceed merging the code. Once the fix is verified, we can easily deal with the merge conflict and we have done it before.

On our end, we need to find a way to address this both from our internal code and AOSP as I have suggested when we received your patch. We will keep communicating with you through the bug especially we would like to make sure we can reproduce this. But please expect delay due to the current situation.

Thanks,
Peiyong

Patch-set: 6
Reviewer: Gerrit User 1267775 <1267775@85c56323-6fa9-3386-8a01-6480fb634889>