Update patch set 2
Patch Set 2:

Thanks for this. That caught me out elsewhere too. It seems as though the absence of the //apex_available:platform causes a difference in the build that causes problems at runtime but is not detected by Soong. That is a serious hole that needs fixing.

Patch-set: 2
Reviewer: Gerrit User 1039338 <1039338@85c56323-6fa9-3386-8a01-6480fb634889>