Remove extra getService calls.
I have not been able to reproduce this race since various servicemanager
fixes have been made in the course of spivack@ adding aidl_lazy_test.
With this, if there is an issue, we will get notified.

This is trading risk:
accepting: potential hang during Android S development
declining: living w/ 1s of delay in some lazy usecases

Since I cannot reproduce a hang, I think this change will be
accepting minimal risk, and in the case we hit that risk, we will
be able to improve the reliability further.

Bug: 140310064
Test: aidl_lazy_test
Change-Id: Ie3d114cc06ab10fd554fac71af35f5bb09bf6c1f