Modify surfaceflnger transaction sync flow
Surfaceflinger use syncpoint to sync two layer's buffer and transaction.
But it doesn't consider some case, so transaction sometimes will miss
buffer.
In CtsViewTestCases, test will check transaction of SurfaceView is matched
to CTS UI buffer or not, if not matched will throw error to CTS.

This patch fix some issue in this flow.
1.Fence signaled check is not align in whole flow
2.Transaction checked by a mutable frame number
3.Some transaction will be lost, that cause sync fail
4.Acquire buffer frame number is not align to transaction result.

Bug: 154893406
Change-Id: I266f166d671bef9b8aaf07caff7ee819be61b106
Test: run CtsViewTestCases about 30 times