Update patch set 1
Patch Set 1:

Kishore,
I don't think Ray meant a new CL had to be uploaded with the change.
It was just a note on why the two new macros aren't used in the current CL for completeness.
Please abandon this and make this change in base CL itself.

Patch-set: 1
Reviewer: Gerrit User 1047003 <1047003@85c56323-6fa9-3386-8a01-6480fb634889>