Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Kishore,
> I don't think Ray meant a new CL had to be uploaded with the change.
> It was just a note on why the two new macros aren't used in the current CL for completeness.
> Please abandon this and make this change in base CL itself.

i was thinking it would be an update to that CL (e.g. another patchset).
perhaps I phrased it poorly.
my preference would be that we stay with the original CL, another patchset/version of that CL
because it means all of the review comments are maintained.

Patch-set: 1
Reviewer: Gerrit User 1108919 <1108919@85c56323-6fa9-3386-8a01-6480fb634889>