Rust Binder throughput benchmark
A direct port of the binderThroughputTest from the C++ binder tests,
built on top of the Binder Rust crate.

This benchmark is significantly slower than the C++ version because the
Rust crate is built on top of the NDK Binder library which forces heap
allocation of Parcels inside the benchmark timing loop.

Test: m rustBinderThroughputBenchmark
Test: adb shell rustBinderThroughputBenchmark -m 1000 -i 100000
Change-Id: I22d91fb397bfff634606f285e6899cd812a62d24