Add Rust Binder throughput benchmark
This is a direct port of the binderThroughputTest from the C++ binder
tests.

We use bindings to C++ to measure time the exact same way in the Rust
binder throughput benchmark as in the C++ version. Without this, the
results were incomparable because Rust time measurement via
std::time::Instant added significant overhead inside the timed interval
compared to the C++ timing measurement in the baseline benchmark.

Test: m rustBinderThroughputTest
Test: adb shell rustBinderThroughputTest -m 1000 -i 100000
Change-Id: I2565c1e870dd4886b8c0d1fcf41480165406fded