Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> > Patch Set 2: Code-Review-2
> > 
> > I am not sure I follow. Some libs like libs/gui and libs/ui don't guaranteed to be backward compatible hence I am not sure why these two specifically contribute to mainline modules. Could you please help me understand more?
> 
> For example, in libs/gui/Android.bp, libgui_headers is used by com.android.bluetooth.updatable, com.android.media, and com.android.media.swcodec. And libgui_bufferqueue_static is used by com.android.media.swcodec. 
> 
> Those modules which are part of mainline modules should be developed with awareness of this backward compatibility requirement. Especially for Q-targeting modules. For more context: https://groups.google.com/a/google.com/g/android-aml-apex/c/r4FsmEtO-Jo/m/xN9hIsyTBwAJ
> 
> Adding Dario for better answer.

What Jooyung said.

com.android.media and com.android.media.swcodec are modules which have launched in Q so there is an expectation that they only depend on libraries that can support in Q. The idea is that an artifact coming from a tree based on R or later is supposed to work when installed on a Q device.

I understand this might be a surprise, but this is exactly the reason why we are adding an explicit min_sdk_version argument to the target -- first and foremost so the owners of their code knows what they have been signed up for so they can work with module authors to set expectations regarding compatibility.

Adding marcone@ and josecubero@ from the Media team. You should discuss with them if you have concerns being used in the media modules.

Patch-set: 3
CC: Gerrit User 1001525 <1001525@85c56323-6fa9-3386-8a01-6480fb634889>
CC: Gerrit User 1444590 <1444590@85c56323-6fa9-3386-8a01-6480fb634889>