Update patch set 11
Patch Set 11:

Something I just realized and is worth considering... The binder NDK library has a mostly complete C interface to binder. I didn't realize that existed when I did the C++ opaque bindings, but I don't think it would be too hard to swap out the C++ shims in BinderBindings.cpp for that library. Just a question of whether the rust crate should depend on it.

Patch-set: 11