Update patch set 3
Patch Set 3:

+Pawin to check the hidl interface change.

I think I need to understand if the hidl change is necessary though - In general would be cautious about updating the hidl interface as we would need to support that interface for a long time, as with all of our hidls.

Can we do this in ImageWriter#queueInputImage without adding the callback? prevOwner should have a reference to its corresponding Surface, so we should be able to release the buffer reference from the surface there.

Patch-set: 3
Reviewer: Gerrit User 1105338 <1105338@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1470474 <1470474@85c56323-6fa9-3386-8a01-6480fb634889>