Update patch set 2
Patch Set 2:

> > This test looks like a benchmark. Shouldn't it be in benches directory?
> 
> Certainly could be. This is a direct port for comparison with binderThroughputTest, which is a cc_test in binder/tests/. I'm happy to move it, although we don't have a specific benchmark module type in soong for Rust, do we?

We should but not have a rust_bench module type yet.
According to past experience, I would expect another module type, not a rust_test variant.
I think we need to discuss more off-line to design this module type and required library/code-generator supports.
Until then, does it work for you to keep the source in benches directory and build it as a rust_binary?

Patch-set: 2