Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> This is the first Rust module I see defined in frameworks/native.
> It looks like following cargo package structure and could be published as 'binder' at crates.io later.
> If that's the case, it should be in external/rust/crates/binder.

I'm not sure this will ever be appropriate for publishing to crates.io. It depends on the platform libbinder and should be maintained alongside it.

The cargo.toml is just to make editors happier and give us a temporary way to regenerate bindgen bindings until we have soong support for that.

Patch-set: 11