Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> +Pawin to check the hidl interface change.
> 
> I think I need to understand if the hidl change is necessary though - In general would be cautious about updating the hidl interface as we would need to support that interface for a long time, as with all of our hidls.
> 
> Can we do this in ImageWriter#queueInputImage without adding the callback? prevOwner should have a reference to its corresponding Surface, so we should be able to release the buffer reference from the surface there.

 Glad to hear from you.

     We do need a callback onBufferDetached to release the GraphicBuffer when Consumer detached the buffer.
     And we need to change the hidl​ to do IPC, since the BufferQueueCore and the Surface may in different process.

the detail is following:
When the Consumer detached the buffer, the buffer cached in Surface is useless, we can release it in time.
You may think that is not a big thing, next time we dequeueBuffer, it will return BUFFER_NEEDS_REALLOCATION, the old useless buffer will be released,that is ok.

But our calling process is as follows：
Since postProcessing is slower by picture producer. We need create BufferQueue1(maxBufferCount=34) to cache the pictures for BufferQueue2 consumming  the calling process as graph shown below.
When we do burst snapshot(speed and continue take 100 pictures at one time), The  cached GraphicBuffer's count in BufferQueue1 Producer will arrive at the maxBufferCount (it‘s 34 in our scene)。But later, We never do burst snapshot, Just do take normal picture(take one picture one time),  never destroy the the Surface, either.
one cached GraphicBuffer in BufferQueue1 will be released by taking normal picture. But the other 33 cached buffers will never be released in time. In fact, the cached GraphicBuffers is detached from Consumer,We will never use it again, and the <sp>reference of the 33 GraphicBuffers in BufferQueueCore and BufferItemConsumer have been released, But Surface still hold the <sp>reference of the 33 GraphicBuffers​.
So we add a onBufferDetached callback method to release the cached buffers.the 33 cached buffers will consumer 600M memory, so release it is necessary.

attach:
https://issuetracker.google.com/issues/147608804​

Patch-set: 3