Update patch set 11
Patch Set 11:

(24 comments)

Thanks for the reviews. There are still a few questions in parcel that I'll leave for Daniel as I don't know the answers.

Sorry for all the commented out bits, this was mostly either C++ APIs that we are not exposing or functionality that wasn't quite ready yet. I can rebase more recent commits into here now which should address the bulk of this, and then we should have a better starting point for further discussion.

Patch-set: 11
CC: Gerrit User 1687565 <1687565@85c56323-6fa9-3386-8a01-6480fb634889>