Update patch set 6
Patch Set 6:

After making this benchmark measure time in the exact same way as the C++ baseline and making Parcel stack-allocated in the binder crate, results are much better:

Rust:
    iterations per sec: 82741.36300972568
    collecting results
    killing workers
    average: 0.023647176906500002ms worst: 7.095261ms best: 0.01125ms
    50%: 0.035156 90%: 0.035156 95%: 0.035156 99%: 0.246092

C++:
    iterations per sec: 83147.1
    collecting results
    killing workers
    average:0.0235957ms worst:7.49854ms best:0.01125ms
    50%: 0.035156 90%: 0.035156 95%: 0.035156 99%: 0.246092

This is effectively no difference in the average case. The overall iterations/s is computed using the overall benchmark time (and measured with Rust time measurement not C++), which has some immaterial differences outside of the transaction measurements.

Patch-set: 6