Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> By the way, is there a reason this change is needed in AOSP? It seems quite device-specific.

hi Pawin
Thanks for your reply.
I think it's needed in AOSP, it's not device-specific. all of use models like this will have this issue.It's just the obvious exposure of camera scene and detachBuffer means to detach, but actual detach effect not achieved，and cause the invalid use of large amount of memory.

Patch-set: 5