Update patch set 1
Patch Set 1:

On a walleye device we're looking at about a 25% reduction in throughput from the baseline. This is almost entirely due to the extra heap allocations of the parcels, and the fact that the NDK API forces us to re-initialize a new parcel for each transaction rather than reusing the same parcel. This puts both the parcel allocations and initialization into the timing loop.

Baseline C++ version:
    iterations per sec: 74859.6
    collecting results
    killing workers
    average:0.0263308ms worst:7.94448ms best:0.01125ms
    50%: 0.035156 90%: 0.035156 95%: 0.035156 99%: 0.246092

Rust version:
    iterations per sec: 56531.664491733696
    collecting results
    killing workers
    average: 0.034784296611ms worst: 3.60224ms best: 0.015677ms
    50%: 0.035156 90%: 0.035156 95%: 0.105468 99%: 0.316404

Patch-set: 1
Reviewer: Gerrit User 1416371 <1416371@85c56323-6fa9-3386-8a01-6480fb634889>
CC: Gerrit User 1052399 <1052399@85c56323-6fa9-3386-8a01-6480fb634889>
CC: Gerrit User 1068404 <1068404@85c56323-6fa9-3386-8a01-6480fb634889>